requirements.txt
gradio>=3.0
ultralytics>=8.0
opencv-python
pillow
numpy